requirements.txt
sphinx==1.8.4
confluence-py==1.0.1
nbsphinx==0.4.2
wheel>=0.23.0
boto>=2.39.0
python-dotenv>=0.5.1
jupyter==1.0.0 
altair>=3.2.0
bump2version>=0.5.10
dill>=0.2.5
htmlark==1.0.0
numpy>=1.10.0
pandas>=0.17.1
pillow>=6.1.0
plotly_express>=0.1.8
python-language-server==0.24
pyyaml>=3.12
screeninfo>=0.4
scikit-learn>=0.17.0
scipy>=0.16.0
seaborn>=0.9.0
vega>=2.4.0
selenium>=3.141.0
streamlit==0.49.0
